requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pycryptodome==3.23.0
Pygments==2.19.2
requests==2.32.5
rich==14.1.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
